requirements.txt
alchemy==20.5
asgiref==3.11.0
attrs==25.4.0
certifi==2025.10.5
chardet==3.0.4
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.22.2
drf-spectacular==0.29.0
filelock==3.0.12
greenlet==3.2.4
idna==2.8
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lockfile==0.12.2
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-daemon==2.2.4
PyYAML==6.0.3
referencing==0.37.0
requests==2.22.0
rpds-py==0.30.0
setuptools==80.9.0
SQLAlchemy==2.0.44
sqlparse==0.5.4
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.25.11
